Updated driver visitor hierarchy
