+++ src/main/java/edu/kis/powp/jobs2d/drivers/monitoring/DriverMonitorDecorator.java
@@ -1,16 +1,15 @@
 package edu.kis.powp.jobs2d.drivers.monitoring;
 
+import edu.kis.powp.jobs2d.drivers.AbstractDecorator;
 import edu.kis.powp.jobs2d.drivers.VisitableJob2dDriver;
-import edu.kis.powp.jobs2d.drivers.visitors.DriverVisitor;
 
-public class DriverMonitorDecorator implements VisitableJob2dDriver {
-    private final VisitableJob2dDriver driver;
+public class DriverMonitorDecorator extends AbstractDecorator {
     private final DriverUsageMonitor monitor;
     private final DriverMonitor outputMonitor;
 
     public DriverMonitorDecorator(VisitableJob2dDriver driver, DriverUsageMonitor monitor,
             DriverMonitor outputMonitor) {
-        this.driver = driver;
+        super(driver);
         this.monitor = monitor;
         this.outputMonitor = outputMonitor;
     }
@@ -27,10 +26,5 @@ public class DriverMonitorDecorator implements VisitableJob2dDriver {
         monitor.recordOperationMove(x, y);
         driver.operateTo(x, y);
         outputMonitor.update(x, y, monitor.getHeadDistance(), monitor.getOperationDistance());
-    }
-
-    @Override
-    public void accept(DriverVisitor visitor) {
-        visitor.visit(this);
     }
 }
