+++ src/main/java/edu/kis/powp/jobs2d/transformations/RotateTransformationDecorator.java
@@ -1,7 +1,6 @@
 package edu.kis.powp.jobs2d.transformations;
 
 import edu.kis.powp.jobs2d.drivers.VisitableJob2dDriver;
-import edu.kis.powp.jobs2d.drivers.visitors.DriverVisitor;
 
 public class RotateTransformationDecorator extends TransformationDecorator {
     private final double cos, sin;
@@ -19,11 +18,6 @@ public class RotateTransformationDecorator extends TransformationDecorator {
                 (int) Math.round(x * cos - y * sin),
                 (int) Math.round(x * sin + y * cos)
         };
-    }
-
-    @Override
-    public void accept(DriverVisitor visitor) {
-        visitor.visit(this);
     }
 
 }
