+++ src/main/java/edu/kis/powp/jobs2d/transformations/ScaleTransformationDecorator.java
@@ -1,7 +1,6 @@
 package edu.kis.powp.jobs2d.transformations;
 
 import edu.kis.powp.jobs2d.drivers.VisitableJob2dDriver;
-import edu.kis.powp.jobs2d.drivers.visitors.DriverVisitor;
 
 public class ScaleTransformationDecorator extends TransformationDecorator {
     private final double scaleX;
@@ -18,10 +17,5 @@ public class ScaleTransformationDecorator extends TransformationDecorator {
         int newX = (int) Math.round(scaleX * x);
         int newY = (int) Math.round(scaleY * y);
         return new int[] { newX, newY };
-    }
-
-    @Override
-    public void accept(DriverVisitor visitor){
-        visitor.visit(this);
     }
 }
